requirements.txt
flet
sqlalchemy
python-dotenv
requests
google-generativeai
certifi